init TabBar
