+++ src/styles/index.css
@@ -18,8 +18,8 @@
 @import "./workbench/editor/editor.css";
 @import "./workbench/editor/titleBar/titleBarView.css";
 @import "./workbench/editor/titleBar/functionBarView.css";
-@import "./workbench/editor/titleBar/tabBarView.css";
 @import "./workbench/editor/titleBar/windowBarView.css";
+@import "./workbench/editor/tabBar/tabBarView.css";
 @import "./workbench/editor/markdown/markdownView.css";
 @import "./workbench/editor/markdown/custom-toastui-style.css";
